Move form code to form.tsx from route.tsx
